update changeset and improved pre-action hook with bundler blocking code
