+++ packages/sku/src/program/commands/start-ssr/start-ssr.action.ts
@@ -6,18 +6,11 @@ import { webpackStartSsrHandler } from '@/program/commands/start-ssr/webpack-sta
 export const startSsrAction = async ({
   stats,
   skuContext,
-  experimentalBundler,
 }: {
   stats: StatsChoices;
   skuContext: SkuContext;
-  experimentalBundler: boolean;
 }) => {
-  if (skuContext.bundler === 'vite' && !experimentalBundler) {
-    throw new Error(
-      'The `vite` bundler is experimental. If you want to use it please use the `--experimental-bundler` flag.',
-    );
-  }
-  if (skuContext.bundler === 'vite' && experimentalBundler) {
+  if (skuContext.bundler === 'vite') {
     viteStartSsrHandler(skuContext);
   } else {
     webpackStartSsrHandler({ stats, skuContext });
