+++ packages/sku/src/program/commands/start/start.command.ts
@@ -1,6 +1,5 @@
 import { Command } from 'commander';
 import { startAction } from './start.action.js';
-import { experimentalBundlerOption } from '@/program/options/expirementalBundler/experimentalBundler.option.js';
 import { statsOption } from '@/program/options/stats/stats.option.js';
 
 const startCommand = new Command('start');
@@ -9,7 +8,6 @@ startCommand
   .description(
     'Start the sku development server for a statically-rendered application.',
   )
-  .addOption(experimentalBundlerOption)
   .addOption(statsOption)
   .action(startAction);
