+++ packages/sku/src/program/index.ts
@@ -7,8 +7,8 @@ import { debugOption } from './options/debug/debug.option.js';
 import { configOption } from './options/config/config.option.js';
 import { environmentOption } from './options/environment/environment.option.js';
 import { initDebug } from '@/utils/debug.js';
-import { getSkuContext } from '@/context/createSkuContext.js';
-import { initializeTelemetry } from '@/services/telemetry/index.js';
+import { experimentalBundlerOption } from '@/program/options/expirementalBundler/experimentalBundler.option.js';
+import { preActionHook } from '@/program/hooks/preAction/preAction.hook.js';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
@@ -26,18 +26,13 @@ program
   .addOption(environmentOption)
   .addOption(configOption)
   .addOption(debugOption)
+  .addOption(experimentalBundlerOption)
   .on('option:debug', () => {
     if (program.opts()?.debug) {
       initDebug();
     }
   })
-  .hook('preAction', (_thisCommand, actionCommand) => {
-    const skuContext = getSkuContext({
-      configPath: program.opts()?.config,
-    });
-    initializeTelemetry(skuContext);
-    actionCommand.setOptionValue('skuContext', skuContext);
-  });
+  .hook('preAction', preActionHook);
 
 for (const command of commands) {
   program.addCommand(command);
